+++ MatrixKit/Models/Account/MXKAccount.m
@@ -129,6 +129,8 @@ static MXKAccountOnCertificateChange _onCertificateChangeBlock;
         
         // Refresh device information
         [self loadDeviceInformation:nil failure:nil];
+        
+        [self registerIdentityServiceDidChangeAccessTokenNotification];
     }
     return self;
 }
@@ -156,16 +158,20 @@ static MXKAccountOnCertificateChange _onCertificateChangeBlock;
         NSString *userId = [coder decodeObjectForKey:@"userid"];
         NSString *accessToken = [coder decodeObjectForKey:@"accesstoken"];
         _identityServerURL = [coder decodeObjectForKey:@"identityserverurl"];
+        NSString *identityServerAccessToken = [coder decodeObjectForKey:@"identityserveraccesstoken"];
         
         mxCredentials = [[MXCredentials alloc] initWithHomeServer:homeServerURL
                                                            userId:userId
                                                       accessToken:accessToken];
 
         mxCredentials.identityServer = _identityServerURL;
+        mxCredentials.identityServerAccessToken = identityServerAccessToken;
         mxCredentials.deviceId = [coder decodeObjectForKey:@"deviceId"];  
         mxCredentials.allowedCertificate = [coder decodeObjectForKey:@"allowedCertificate"];
         
         [self prepareRESTClient];
+        
+        [self registerIdentityServiceDidChangeAccessTokenNotification];
 
         if ([coder decodeObjectForKey:@"threePIDs"])
         {
@@ -212,6 +218,7 @@ static MXKAccountOnCertificateChange _onCertificateChangeBlock;
     [coder encodeObject:mxCredentials.homeServer forKey:@"homeserverurl"];
     [coder encodeObject:mxCredentials.userId forKey:@"userid"];
     [coder encodeObject:mxCredentials.accessToken forKey:@"accesstoken"];
+    [coder encodeObject:mxCredentials.identityServerAccessToken forKey:@"identityserveraccesstoken"];
 
     if (mxCredentials.deviceId)
     {
@@ -269,11 +276,18 @@ static MXKAccountOnCertificateChange _onCertificateChangeBlock;
         _identityServerURL = identityServerURL;
         // Update the current restClient
         [mxRestClient setIdentityServer:identityServerURL];
+        
+        mxCredentials.identityServer = identityServerURL;
+        
+        // Update identity service
+        MXIdentityService *identityService = [[MXIdentityService alloc] initWithCredentials:mxCredentials andHomeserverRestClient:mxRestClient];
+        mxSession.identityService = identityService;
     }
     else
     {
-        _identityServerURL = nil;        
+        _identityServerURL = nil;
         [mxRestClient setIdentityServer:nil];
+        mxSession.identityService = nil;
     }
     
     // Archive updated field
@@ -1963,6 +1977,30 @@ static MXKAccountOnCertificateChange _onCertificateChangeBlock;
             // Should never happen
             completion(NO);
         }];
+    }
+}
+
+- (void)identityService:(MXIdentityService *)identityService didUpdateAccessToken:(NSString *)accessToken
+{
+    mxCredentials.identityServerAccessToken = accessToken;
+}
+
+- (void)registerIdentityServiceDidChangeAccessTokenNotification
+{
+    [[NSNotificationCenter defaultCenter] addObserver:self selector:@selector(handleIdentityServiceDidChangeAccessTokenNotification:) name:MXIdentityServiceDidChangeAccessTokenNotification object:nil];
+}
+
+- (void)handleIdentityServiceDidChangeAccessTokenNotification:(NSNotification*)notification
+{
+    NSDictionary *userInfo = notification.userInfo;
+    
+    NSString *userId = userInfo[MXIdentityServiceNotificationUserIdKey];
+    NSString *identityServer = userInfo[MXIdentityServiceNotificationIdentityServerKey];
+    NSString *accessToken = userInfo[MXIdentityServiceNotificationAccessTokenKey];
+    
+    if (userId && identityServer && accessToken && [mxCredentials.identityServer isEqualToString:identityServer])
+    {
+        mxCredentials.identityServerAccessToken = accessToken;
     }
 }